Toimipisteet UI
Toimipisteet käyttöliittymän muokkausta.  Lisätyt tallenna ja peruuta napit, ja niiden toiminnallisuus

+++ mokkisofta/Toimipisteet.cs
@@ -13,6 +13,8 @@ namespace mokkisofta
     public partial class Toimipisteet : Form
     { 
         Sql S = new Sql();
+        private bool btnLisaaPainettu = false;
+        private bool btnMuokkaaPainettu = false;
         public Toimipisteet()
         {
             InitializeComponent();
@@ -21,6 +23,9 @@ namespace mokkisofta
             DgwToimipisteet.DataSource = S.ShowInGridView("SELECT toimipiste_id AS Id, nimi AS Nimi, lahiosoite AS Osoite, postitoimipaikka AS paikkakunta, postinro AS Postinumero, email AS Sähköposti, puhelinnro AS Puhelin FROM Toimipiste");
             S.Close();
 
+            this.Controls.OfType<TextBox>().ToList().ForEach(t => t.Enabled = false); // Ohjelman käynnistyessä tekstikenttiin ei voi syöttää tietoa.
+            btnTpTallenna.Enabled = false;
+            btnTpPeruuta.Enabled = false;
             /*
             // Määritetään datagridview niin, että koko rivi tulee aina valituksi.
             DgwToimipisteet.SelectionMode = DataGridViewSelectionMode.FullRowSelect;
@@ -43,41 +48,35 @@ namespace mokkisofta
 
         private void BtnTpLisaa_Click(object sender, EventArgs e)
         {
-            S.Connect();
-
-            string tpNimi = txbTpNimi.Text;
-            string tpOsoite = txbTpOsoite.Text;
-            string tpPtoimipaikka = txbTpPtoimipaikka.Text;
-            string tpPostinro = txbTpPostinumero.Text;
-            string tpSposti = txbTpSposti.Text;
-            string tpPuhnro = txbTpPuhnro.Text;
-            if (this.Controls.OfType<TextBox>().Any(t => string.IsNullOrEmpty(t.Text)))
-            {
-                MessageBox.Show("Syötä tiedot kaikkiin tekstikenttiin!");
-            }
-            else
-            {
-                string tpLisays = $"INSERT INTO Toimipiste (nimi, lahiosoite, postitoimipaikka, postinro, email, puhelinnro) VALUES ('{tpNimi}', '{tpOsoite}', '{tpPtoimipaikka}', '{tpPostinro}', '{tpSposti}', '{tpPuhnro}')";
-
-                S.Query(tpLisays);
-                DgwToimipisteet.DataSource = S.ShowInGridView("SELECT toimipiste_id AS Id, nimi AS Nimi, lahiosoite AS Osoite, postitoimipaikka AS paikkakunta, postinro AS Postinumero, email AS Sähköposti, puhelinnro AS Puhelin FROM Toimipiste");
-                
-            }
-            S.Close();
+            this.Controls.OfType<TextBox>().ToList().ForEach(t => t.Enabled = true);
+            btnLisaaPainettu = true;
+            btnTpLisaa.Enabled = false;
+            btnTpMuokkaa.Enabled = false;
+            btnTpPoista.Enabled = false;
+            btnTpTallenna.Enabled = true;
+            btnTpPeruuta.Enabled = true;
         }
 
         private void BtnTpPoista_Click(object sender, EventArgs e)
         {
             S.Connect();
-            if (DgwToimipisteet.CurrentCell != null)
+            DialogResult kysely = MessageBox.Show("Haluatko varmasti poistaa valitun rivin?", "Poistetaanko?", MessageBoxButtons.OKCancel);
+            if (kysely == DialogResult.OK)
             {
-                int rowIndex = DgwToimipisteet.CurrentCell.RowIndex;
-                string valittuRivi = DgwToimipisteet.Rows[rowIndex].Cells["Id"].Value.ToString();
-                string tpPoisto = $"DELETE FROM Toimipiste WHERE toimipiste_id='{valittuRivi}'";
+                if (DgwToimipisteet.CurrentCell != null)
+                {
+                    int rowIndex = DgwToimipisteet.CurrentCell.RowIndex;
+                    string valittuRivi = DgwToimipisteet.Rows[rowIndex].Cells["Id"].Value.ToString();
+                    string tpPoisto = $"DELETE FROM Toimipiste WHERE toimipiste_id='{valittuRivi}'";
 
-                S.Query(tpPoisto);
-                DgwToimipisteet.DataSource = S.ShowInGridView("SELECT toimipiste_id AS Id, nimi AS Nimi, lahiosoite AS Osoite, postitoimipaikka AS paikkakunta, postinro AS Postinumero, email AS Sähköposti, puhelinnro AS Puhelin FROM Toimipiste");
-                S.Close();
+                    S.Query(tpPoisto);
+                    DgwToimipisteet.DataSource = S.ShowInGridView("SELECT toimipiste_id AS Id, nimi AS Nimi, lahiosoite AS Osoite, postitoimipaikka AS paikkakunta, postinro AS Postinumero, email AS Sähköposti, puhelinnro AS Puhelin FROM Toimipiste");
+                    S.Close();
+                }
+            }
+            if (kysely == DialogResult.Cancel)
+            {
+
             }
         }
 
@@ -98,11 +97,64 @@ namespace mokkisofta
             }
         }
 
-        private void DgwToimipisteet_CellContentClick(object sender, DataGridViewCellEventArgs e)
-        {
 
+        private void BtnTpTallenna_Click(object sender, EventArgs e)
+        {
+            /* Ehtolauseella tarkastetaan halutaanko lisätä vai muokata tietoja tietokantaan. 
+             * Muuttujina ovat btnLisaaPainettu ja btnMuokkaaPainettu. Toiminnon jälkeen kyseiset muuttujat saavat arvon false.
+             * Toiminnon suoritettua kaikki tekstikentät nollataan ja painonapit ovat jälleen käytettävissä.
+             */
+            S.Connect();
+            if (btnLisaaPainettu == true)
+            {
+                string tpNimi = txbTpNimi.Text;
+                string tpOsoite = txbTpOsoite.Text;
+                string tpPtoimipaikka = txbTpPtoimipaikka.Text;
+                string tpPostinro = txbTpPostinumero.Text;
+                string tpSposti = txbTpSposti.Text;
+                string tpPuhnro = txbTpPuhnro.Text;
+                if (this.Controls.OfType<TextBox>().Any(t => string.IsNullOrEmpty(t.Text)))
+                {
+                    MessageBox.Show("Syötä tiedot kaikkiin tekstikenttiin!");
+                }
+                else
+                {
+                    string tpLisays = $"INSERT INTO Toimipiste (nimi, lahiosoite, postitoimipaikka, postinro, email, puhelinnro) VALUES ('{tpNimi}', '{tpOsoite}', '{tpPtoimipaikka}', '{tpPostinro}', '{tpSposti}', '{tpPuhnro}')";
+
+                    S.Query(tpLisays);
+                    DgwToimipisteet.DataSource = S.ShowInGridView("SELECT toimipiste_id AS Id, nimi AS Nimi, lahiosoite AS Osoite, postitoimipaikka AS paikkakunta, postinro AS Postinumero, email AS Sähköposti, puhelinnro AS Puhelin FROM Toimipiste");
+                    btnLisaaPainettu = false;
+                    btnMuokkaaPainettu = false;
+                    this.Controls.OfType<TextBox>().ToList().ForEach(t => t.Text = string.Empty);
+                    this.Controls.OfType<TextBox>().ToList().ForEach(t => t.Enabled = false);
+                    btnTpLisaa.Enabled = true;
+                    btnTpMuokkaa.Enabled = true;
+                    btnTpPoista.Enabled = true;
+                    btnTpTallenna.Enabled = false; ;
+                    btnTpPeruuta.Enabled = false; ;
+                }
+            }
+            else if (btnMuokkaaPainettu == true)
+            {
+
+            }
+
+            S.Close();
         }
 
+        private void BtnTpPeruuta_Click(object sender, EventArgs e)
+        {
+            this.Controls.OfType<TextBox>().ToList().ForEach(t => t.Text = string.Empty);
+            this.Controls.OfType<TextBox>().ToList().ForEach(t => t.Enabled = false);
+            btnTpLisaa.Enabled = true;
+            btnTpMuokkaa.Enabled = true;
+            btnTpPoista.Enabled = true;
+            btnLisaaPainettu = false;
+            btnMuokkaaPainettu = false;
+            btnTpTallenna.Enabled = false; ;
+            btnTpPeruuta.Enabled = false; ;
+
+        }
 
 
         /*
